Only support grenade airtime for smokes.
This fixes some problems, namely:
- flashes always have an airtime of 1.6 seconds. Not that useful.
- molotov/inc event being hooked doesn't have an entity param.

There is an event for inferno_start that has an entity, but it still
isn't that useful and hooking off that may not be reliable. We need
the guarantee that the event is ALWAYS called. What if the moltov
is put out by a smoke and inferno_start is never sent?